9.0.0:使用XToolsUni，新支持MacOS-Arm和NILRT。字符编码全平台统一为UTF8。新增Vars::spinEnter/Leave方法，SpinLocked类。优化as/is,移除cast。新增Safe类实现重定向多线程安全。新增StringSpan和StringAppender。String和StringSpan都基于StringCommon，并增加toNumber方法。split和subString都返回StringSpan类型。移除String::updateLength。优化字符串处理效率。改良Enum框架，实现有效性判断和转字符串。Matrix模板化。使用C++标准库中的function。规范注释[]标签。移除SpadasSupport。优化msys控制台IO。ArrayElem改为class，ListElem不继承Object，ImagePointer改为class。多个类的clear方法改为set，避免混淆。Stream固定使用自旋锁。移除OpenCV兼容。math::round/floor/ceil返回Long。TimeWithMS.substrct改为minus。VideoDeviceData.hasPreview/preview合并

+++ Spadas/math.cpp
@@ -1,6 +1,5 @@
 ﻿
 #include "spadas.h"
-
 #include <stdlib.h>
 #include <math.h>
 
@@ -285,7 +284,7 @@ Array<Float> spadas::math::random(UInt size)
 	srand(timePtr[0] + timePtr[1]);
 
 	Array<Float> out(size);
-	Float k = 1.0f / RAND_MAX;
+	Float k = 1.0f / (Float)RAND_MAX;
 	for (UInt i = 0; i < size; i++)
 	{
 		out[i] = (Float)rand() * k;
@@ -301,7 +300,7 @@ Array<Double> spadas::math::randomD(UInt size)
 	srand(timePtr[0] + timePtr[1]);
 
 	Array<Double> out(size);
-	Double k = 1.0f / RAND_MAX;
+	Double k = 1.0f / (Float)RAND_MAX;
 	for (UInt i = 0; i < size; i++)
 	{
 		out[i] = (Double)rand() * k;
